refactor: extract SetActionButton component from set display
Extract inline action button JSX and logic into dedicated SetActionButton component for better separation of concerns.
